Relicense this flake from GPL-3 -> MIT to allow use in nixpkgs
Originally, I'd used the GPL-3 to follow suit with blackbox itself.
Upon further consideration I think it makes a lot more sense for this
flake to be licensed under MIT.

In particular, this should enable nixpkgs to use this code when
upstreaming support for the blackbox package.

Fwiw, I'm the copyright holder of this flake and the only contributor to
this repository so far, and the flake itself is quite new, so hopefully
this should not cause any issues.